Moved class pleyer to new file
